Деструктуризация компонента на составляющие
